event(event): event

+++ code/modules/reagents/reagent_dispenser.dm
@@ -1,136 +1,145 @@
 
+
 /obj/structure/reagent_dispensers
 	name = "Dispenser"
 	desc = "..."
 	icon = 'icons/obj/objects.dmi'
 	icon_state = "watertank"
-	density = 1
-	anchored = 0
-	pull_sound = "pull_machine"
+	density = TRUE
+	anchored = FALSE
 
-	var/initial_capacity = 1000
-	var/initial_reagent_types  // A list of reagents and their ratio relative the initial capacity. list(/datum/reagent/water = 0.5) would fill the dispenser halfway to capacity.
 	var/amount_per_transfer_from_this = 10
-	var/possible_transfer_amounts = "10;25;50;100;500"
+	var/possible_transfer_amounts = list(10,25,50,100)
 
 	attackby(obj/item/weapon/W as obj, mob/user as mob)
 		return
 
-/obj/structure/reagent_dispensers/New()
-	create_reagents(initial_capacity)
+	New()
+		var/datum/reagents/R = new/datum/reagents(1000)
+		reagents = R
+		R.my_atom = src
+		if (!possible_transfer_amounts)
+			verbs -= /obj/structure/reagent_dispensers/verb/set_APTFT
+		..()
 
-	if (!possible_transfer_amounts)
-		src.verbs -= /obj/structure/reagent_dispensers/verb/set_APTFT
-
-	for(var/reagent_type in initial_reagent_types)
-		var/reagent_ratio = initial_reagent_types[reagent_type]
-		reagents.add_reagent(reagent_type, reagent_ratio * initial_capacity)
-
-	..()
-
-/obj/structure/reagent_dispensers/examine(mob/user)
-	if(!..(user, 2))
-		return
-	to_chat(user, "<span class='notice'>It contains:</span>")
-	if(reagents && reagents.reagent_list.len)
-		for(var/datum/reagent/R in reagents.reagent_list)
-			to_chat(user, "<span class='notice'>[R.volume] units of [R.name]</span>")
-	else
-		to_chat(user, "<span class='notice'>Nothing.</span>")
-
-/obj/structure/reagent_dispensers/verb/set_APTFT() //set amount_per_transfer_from_this
-	set name = "Set transfer amount"
-	set category = "Object"
-	set src in view(1)
-	var/N = input("Amount per transfer from this:","[src]") as null|anything in cached_number_list_decode(possible_transfer_amounts)
-	if (N)
-		amount_per_transfer_from_this = N
-
-/obj/structure/reagent_dispensers/ex_act(severity)
-	switch(severity)
-		if(1.0)
-			qdel(src)
+	examine(mob/user)
+		if (!..(user, 2))
 			return
-		if(2.0)
-			if (prob(50))
-				new /obj/effect/effect/water(src.loc)
-				qdel(src)
-				return
-		if(3.0)
-			if (prob(5))
-				new /obj/effect/effect/water(src.loc)
-				qdel(src)
-				return
+		user << "<span class = 'notice'>It contains:</span>"
+		if (reagents && reagents.reagent_list.len)
+			for (var/datum/reagent/R in reagents.reagent_list)
+				user << "<span class = 'notice'>[R.volume] units of [R.name]</span>"
 		else
-	return
+			user << "<span class = 'notice'>Nothing.</span>"
 
-/obj/structure/reagent_dispensers/AltClick(mob/user)
-	if(possible_transfer_amounts)
-		if(CanPhysicallyInteract(user))
-			set_APTFT()
-	else
-		return ..()
+	verb/set_APTFT() //set amount_per_transfer_from_this
+		set name = "Set transfer amount"
+		set category = null
+		set src in view(1)
+		var/N = input("Amount per transfer from this:","[src]") as null|anything in possible_transfer_amounts
+		if (N)
+			amount_per_transfer_from_this = N
+
+	ex_act(severity)
+		switch(severity)
+			if (1.0)
+				qdel(src)
+				return
+			if (2.0)
+				if (prob(50))
+					new /obj/effect/effect/water(loc)
+					qdel(src)
+					return
+			if (3.0)
+				if (prob(5))
+					new /obj/effect/effect/water(loc)
+					qdel(src)
+					return
+			else
+		return
+
+
+
+
+
 
 
 //Dispensers
 /obj/structure/reagent_dispensers/watertank
 	name = "watertank"
-	desc = "A tank containing water."
+	desc = "A watertank. It is used to store high amounts of water."
 	icon = 'icons/obj/objects.dmi'
 	icon_state = "watertank"
 	amount_per_transfer_from_this = 10
-	possible_transfer_amounts = "10;25;50;100"
-	initial_capacity = 50000
-	initial_reagent_types = list(/datum/reagent/water = 1)
-	atom_flags = ATOM_FLAG_CLIMBABLE
+	New()
+		..()
+		reagents.add_reagent("water",500)
+
+/obj/structure/reagent_dispensers/watertank/huge
+	name = "high-volume watertank"
+	desc = "A high-volume watertank. It is used to store HUGE amounts of water."
+	icon_state = "hvwatertank"
+	New()
+		..()
+		reagents.add_reagent("water",500)		//Adds 500 units to the amount, that already is inside. It'll be 1000.
 
 /obj/structure/reagent_dispensers/fueltank
-	name = "fueltank"
-	desc = "A tank containing fuel."
-	icon = 'icons/obj/objects.dmi'
-	icon_state = "weldtank"
-	amount_per_transfer_from_this = 10
-	var/modded = 0
-	var/obj/item/device/assembly_holder/rig = null
-	initial_reagent_types = list(/datum/reagent/fuel = 1)
-	atom_flags = ATOM_FLAG_CLIMBABLE
+	name = "fuel barrel"
+	desc = "A fuel barrel. It is used to store high amounts of fuel."
+	icon = 'icons/obj/barrel.dmi'
+	icon_state = "barrel"
+	amount_per_transfer_from_this = 25
+	var/modded = FALSE
+	var/obj/item/assembly_holder/rig = null
+	New()
+		..()
+		reagents.add_reagent("fuel",500)
+
+/obj/structure/reagent_dispensers/fueltank/huge
+	name = "high-volume fueltank"
+	desc = "A high-volume fueltank. It is used to store HUGE amounts of fuel."
+	icon_state = "hvweldtank"
+	New()
+		..()
+		reagents.add_reagent("fuel",500)		//Adds 500 units to the amount, that already is inside. It'll be 1000.
 
 /obj/structure/reagent_dispensers/fueltank/examine(mob/user)
-	if(!..(user, 2))
+	if (!..(user, 2))
 		return
 	if (modded)
-		to_chat(user, "<span class='warning'>Fuel faucet is wrenched open, leaking the fuel!</span>")
-	if(rig)
-		to_chat(user, "<span class='notice'>There is some kind of device rigged to the tank.</span>")
+		user << "<span class = 'red'>The fuel faucet is wrenched open, leaking the fuel!</span>"
+	if (rig)
+		user << "<span class='notice'>There is some kind of device rigged to the tank.</span>"
+
 /obj/structure/reagent_dispensers/fueltank/attack_hand()
 	if (rig)
 		usr.visible_message("<span class='notice'>\The [usr] begins to detach [rig] from \the [src].</span>", "<span class='notice'>You begin to detach [rig] from \the [src].</span>")
-		if(do_after(usr, 20, src))
+		if (do_after(usr, 20, src))
 			usr.visible_message("<span class='notice'>\The [usr] detaches \the [rig] from \the [src].</span>", "<span class='notice'>You detach [rig] from \the [src]</span>")
 			rig.loc = get_turf(usr)
 			rig = null
-			overlays = new /list()
+			overlays = new/list()
 
 /obj/structure/reagent_dispensers/fueltank/attackby(obj/item/weapon/W as obj, mob/user as mob)
-	src.add_fingerprint(user)
+	add_fingerprint(user)
 	if (istype(W,/obj/item/weapon/wrench))
 		user.visible_message("[user] wrenches [src]'s faucet [modded ? "closed" : "open"].", \
 			"You wrench [src]'s faucet [modded ? "closed" : "open"]")
-		modded = modded ? 0 : 1
+		modded = modded ? FALSE : TRUE
 		if (modded)
 			message_admins("[key_name_admin(user)] opened fueltank at [loc.loc.name] ([loc.x],[loc.y],[loc.z]), leaking fuel. (<A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[loc.x];Y=[loc.y];Z=[loc.z]'>JMP</a>)")
 			log_game("[key_name(user)] opened fueltank at [loc.loc.name] ([loc.x],[loc.y],[loc.z]), leaking fuel.")
 			leak_fuel(amount_per_transfer_from_this)
-	else if (istype(W,/obj/item/device/assembly_holder))
+	if (istype(W,/obj/item/assembly_holder))
 		if (rig)
-			to_chat(user, "<span class='warning'>There is another device in the way.</span>")
+			user << "<span class='warning'>There is another device in the way.</span>"
 			return ..()
 		user.visible_message("\The [user] begins rigging [W] to \the [src].", "You begin rigging [W] to \the [src]")
-		if(do_after(user, 20, src))
-			user.visible_message("<span class='notice'>The [user] rigs [W] to \the [src].</span>", "<span class='notice'>You rig [W] to \the [src].</span>")
+		if (do_after(user, 20, src))
+			user.visible_message("<span class='notice'>The [user] rigs [W] to \the [src].", "<span class = 'notice'> You rig [W] to \the [src].</span>")
 
-			var/obj/item/device/assembly_holder/H = W
-			if (istype(H.a_left,/obj/item/device/assembly/igniter) || istype(H.a_right,/obj/item/device/assembly/igniter))
+			var/obj/item/assembly_holder/H = W
+			if (istype(H.a_left,/obj/item/assembly/igniter) || istype(H.a_right,/obj/item/assembly/igniter))
 				message_admins("[key_name_admin(user)] rigged fueltank at [loc.loc.name] ([loc.x],[loc.y],[loc.z]) for explosion. (<A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[loc.x];Y=[loc.y];Z=[loc.z]'>JMP</a>)")
 				log_game("[key_name(user)] rigged fueltank at [loc.loc.name] ([loc.x],[loc.y],[loc.z]) for explosion.")
 
@@ -143,45 +152,31 @@
 			test.Shift(EAST,6)
 			overlays += test
 
-	else if(isflamesource(W))
-		log_and_message_admins("triggered a fueltank explosion with [W].")
-		user.visible_message("<span class='danger'>[user] puts [W] to [src]!</span>", "<span class='danger'>You put \the [W] to \the [src] and with a moment of lucidity you realize, this might not have been the smartest thing you've ever done.</span>")
-		src.explode()
-
-		return
-
 	return ..()
 
 
-/obj/structure/reagent_dispensers/fueltank/bullet_act(obj/item/projectile/Proj)
-	if(Proj.get_structure_damage())
-		if(istype(Proj.firer))
-			var/turf/turf = get_turf(src)
-			if(turf)
-				var/area/area = turf.loc || "*unknown area*"
-				message_admins("[key_name_admin(Proj.firer)] shot a fueltank in \the [area] ([turf.x],[turf.y],[turf.z]) (<A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[turf.x];Y=[turf.y];Z=[turf.z]'>JMP</a>).")
-				log_game("[key_name(Proj.firer)] shot a fueltank in \the [area] ([turf.x],[turf.y],[turf.z]).")
-			else
-				message_admins("[key_name_admin(Proj.firer)] shot a fueltank outside the world.")
-				log_game("[key_name(Proj.firer)] shot a fueltank outside the world.")
+/obj/structure/reagent_dispensers/fueltank/bullet_act(var/obj/item/projectile/Proj)
+	if (Proj.get_structure_damage())
+		if (Proj.firer)
+			message_admins("[key_name_admin(Proj.firer)] shot fueltank at [loc.loc.name] ([loc.x],[loc.y],[loc.z]) (<A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[loc.x];Y=[loc.y];Z=[loc.z]'>JMP</a>).")
+			log_game("[key_name(Proj.firer)] shot fueltank at [loc.loc.name] ([loc.x],[loc.y],[loc.z]).")
 
-		if(!istype(Proj ,/obj/item/projectile/beam/lastertag) && !istype(Proj ,/obj/item/projectile/beam/practice) )
-			explode()
+		//if (!istype(Proj ,/obj/item/projectile/beam/practice) )
+		explode()
 
 /obj/structure/reagent_dispensers/fueltank/ex_act()
 	explode()
 
 /obj/structure/reagent_dispensers/fueltank/proc/explode()
 	if (reagents.total_volume > 500)
-		explosion(src.loc,1,2,4)
+		explosion(loc,1,2,4,2)
 	else if (reagents.total_volume > 100)
-		explosion(src.loc,0,1,3)
+		explosion(loc,0,1,3,1)
 	else if (reagents.total_volume > 50)
-		explosion(src.loc,-1,1,2)
-	if(src)
-		qdel(src)
+		explosion(loc,-1,1,2,1)
+	if (src) qdel(src)
 
-/obj/structure/reagent_dispensers/fueltank/fire_act(datum/gas_mixture/air, temperature, volume)
+/obj/structure/reagent_dispensers/fueltank/fire_act(temperature)
 	if (modded)
 		explode()
 	else if (temperature > T0C+500)
@@ -193,22 +188,24 @@
 		leak_fuel(amount_per_transfer_from_this/10.0)
 
 /obj/structure/reagent_dispensers/fueltank/proc/leak_fuel(amount)
-	if (reagents.total_volume == 0)
+	if (reagents.total_volume == FALSE)
 		return
 
 	amount = min(amount, reagents.total_volume)
-	reagents.remove_reagent(/datum/reagent/fuel,amount)
-	new /obj/effect/decal/cleanable/liquid_fuel(src.loc, amount,1)
+	reagents.remove_reagent("fuel",amount)
+	new /obj/effect/decal/cleanable/liquid_fuel(loc, amount,1)
 
 /obj/structure/reagent_dispensers/peppertank
 	name = "Pepper Spray Refiller"
-	desc = "Refills pepper spray canisters."
+	desc = "Refill pepper spray canisters."
 	icon = 'icons/obj/objects.dmi'
 	icon_state = "peppertank"
-	anchored = 1
-	density = 0
+	anchored = TRUE
+	density = FALSE
 	amount_per_transfer_from_this = 45
-	initial_reagent_types = list(/datum/reagent/capsaicin/condensed = 1)
+	New()
+		..()
+		reagents.add_reagent("condensedcapsaicin",1000)
 
 
 /obj/structure/reagent_dispensers/water_cooler
@@ -218,21 +215,22 @@
 	icon = 'icons/obj/vending.dmi'
 	icon_state = "water_cooler"
 	possible_transfer_amounts = null
-	anchored = 1
-	initial_capacity = 500
-	initial_reagent_types = list(/datum/reagent/water = 1)
+	anchored = TRUE
+	New()
+		..()
+		reagents.add_reagent("water",500)
 
 /obj/structure/reagent_dispensers/water_cooler/attackby(obj/item/weapon/W as obj, mob/user as mob)
 	if (istype(W,/obj/item/weapon/wrench))
-		src.add_fingerprint(user)
-		if(anchored)
+		add_fingerprint(user)
+		if (anchored)
 			user.visible_message("\The [user] begins unsecuring \the [src] from the floor.", "You start unsecuring \the [src] from the floor.")
 		else
 			user.visible_message("\The [user] begins securing \the [src] to the floor.", "You start securing \the [src] to the floor.")
 
-		if(do_after(user, 20, src))
-			if(!src) return
-			to_chat(user, "<span class='notice'>You [anchored? "un" : ""]secured \the [src]!</span>")
+		if (do_after(user, 20, src))
+			if (!src) return
+			user << "<span class='notice'>You [anchored? "un" : ""]secured \the [src]!</span>"
 			anchored = !anchored
 		return
 	else
@@ -240,12 +238,13 @@
 
 /obj/structure/reagent_dispensers/beerkeg
 	name = "beer keg"
-	desc = "A beer keg."
+	desc = "A beer keg"
 	icon = 'icons/obj/objects.dmi'
 	icon_state = "beertankTEMP"
 	amount_per_transfer_from_this = 10
-	initial_reagent_types = list(/datum/reagent/ethanol/beer = 1)
-	atom_flags = ATOM_FLAG_CLIMBABLE
+	New()
+		..()
+		reagents.add_reagent("beer",1000)
 
 /obj/structure/reagent_dispensers/virusfood
 	name = "Virus Food Dispenser"
@@ -253,8 +252,11 @@
 	icon = 'icons/obj/objects.dmi'
 	icon_state = "virusfoodtank"
 	amount_per_transfer_from_this = 10
-	anchored = 1
-	initial_reagent_types = list(/datum/reagent/nutriment/virus_food = 1)
+	anchored = TRUE
+
+	New()
+		..()
+		reagents.add_reagent("virusfood", 1000)
 
 /obj/structure/reagent_dispensers/acid
 	name = "Sulphuric Acid Dispenser"
@@ -262,5 +264,8 @@
 	icon = 'icons/obj/objects.dmi'
 	icon_state = "acidtank"
 	amount_per_transfer_from_this = 10
-	anchored = 1
-	initial_reagent_types = list(/datum/reagent/acid = 1)
+	anchored = TRUE
+
+	New()
+		..()
+		reagents.add_reagent("sacid", 1000)
